Misc whitespace, deletions, and simplifications.
