+++ src/account-settings/delete-account/index.js
@@ -1,4 +1,4 @@
 export { default } from './DeleteAccount';
 export { default as reducer } from './data/reducers';
 export { default as saga } from './data/sagas';
-export { configure } from './data/service';
+export { configureService } from './data/service';
