+++ src/account-settings/site-language/selectors.js
@@ -1,0 +1,20 @@
+import { createSelector } from 'reselect';
+import { getModuleState } from '../../common/utils';
+
+export const storePath = ['accountSettings', 'siteLanguage'];
+
+const siteLanguageSelector = state => getModuleState(state, storePath);
+
+export const siteLanguageListSelector = createSelector(
+  siteLanguageSelector,
+  siteLanguage => siteLanguage.siteLanguageList,
+);
+
+export const siteLanguageOptionsSelector = createSelector(
+  siteLanguageSelector,
+  siteLanguage =>
+    siteLanguage.siteLanguageList.map(({ code, name }) => ({
+      value: code,
+      label: name,
+    })),
+);
