--- src/site-language/index.js
@@ -1,16 +1,0 @@
-import reducer from './reducers';
-import saga from './sagas';
-import { configureApiService, ApiService } from './service';
-import { storeName, siteLanguageOptionsSelector, siteLanguageListSelector } from './selectors';
-import { fetchSiteLanguages } from './actions';
-
-export {
-  reducer,
-  saga,
-  configureApiService,
-  storeName,
-  siteLanguageListSelector,
-  siteLanguageOptionsSelector,
-  fetchSiteLanguages,
-  ApiService,
-};
